chore: drop poseidonFunc TODO and ts-ignore in merkle/hash.ts
cryptography#7 typed poseidonFunc properly: it now accepts
Uint8Array[] and narrows the return type to Uint8Array via the conditional
PoseidonFuncResult<R, N>. The TODO comment and two @ts-ignore directives
in poseidonHash were placeholders waiting for that fix and are now
redundant.

+++ src/merkle/hash.ts
@@ -8,10 +8,7 @@ import { poseidonFunc } from '@railgun-reloaded/cryptography'
  * @returns 32 bytes hash output
  */
 function poseidonHash (out: Uint8Array, left: Readonly<Uint8Array>, right: Readonly<Uint8Array>) {
-  // TODO: Update poseidonFunc type signature to accept Uint8Array[] instead of any
-  // @ts-ignore
-  const hash = poseidonFunc([left, right])
-  // @ts-ignore
+  const hash = poseidonFunc([left, right]) as Uint8Array
   out.set(hash)
   return out
 }
